tests(misc): add tests to increase coverage(reformat)
